pyproject.toml
[tool.poetry]
name = "telegramparser"
version = "0.1.0"
description = ""
authors = ["Alemkhan <alemkhan.yergaliev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
Telethon = "1.24.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
